user tasklist

+++ app/Http/Controllers/TasksController.php
@@ -23,13 +23,30 @@ class TasksController extends Controller
      */
     public function index()
     {
-        //メッセージ一覧を取得
-        $tasks = Task::all();
+        
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
 
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+            $data = [
+                'user' => $user,
+                'tasks' => $tasks,
+            ];
+        }
+
+        // Welcomeビューでそれらを表示
+        return view('tasks.index', $data);
+
+
+        // //メッセージ一覧を取得
+        // $tasks = Task::all();
+        // // メッセージ一覧ビューでそれを表示
+        // return view('tasks.index', [
+        //     'tasks' => $tasks,
+        // ]);
     }
 
     /**
@@ -61,13 +78,23 @@ class TasksController extends Controller
         ]);
         
         //タスクの作成
-        $task = new Task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+         // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create ([
+            'status' => $request->status,
+            'content' => $request->content,
+        ]);
+
+        return back();
         
-        //トップページにリダイレクト
-        return redirect('/');
+        // $task = new Task;
+        // $task->status = $request->status;
+        // $task->content = $request->content;
+        // $task->save();
+        
+        // //トップページにリダイレクト
+        // return redirect('/');
+        
+        
         // return view('tasks.index', [
         //     'tasks' => $tasks,
         // ]);
@@ -81,13 +108,20 @@ class TasksController extends Controller
      */
     public function show($id)
     {
-        //idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        
-        //メッセージ詳細ビューでそれを表示
-        return view('tasks.show', [
-            'task' => $task,
+             //idの値でメッセージを検索して取得
+            $task = Task::findOrFail($id);
+
+         if (\Auth::id() === $task->user_id) {
+            //メッセージ詳細ビューでそれを表示
+            return view('tasks.show', [
+                'task' => $task,
             ]);
+         }
+         else {
+            // 前のURLへリダイレクトさせる
+            return back();
+        }
+
     }
 
     /**
@@ -98,13 +132,19 @@ class TasksController extends Controller
      */
     public function edit($id)
     {
-        //idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-
-        // メッセージ編集ビューでそれを表示
-        return view('tasks.edit', [
-            'task' => $task,
-         ]);
+            //idの値でメッセージを検索して取得
+            $task = Task::findOrFail($id);
+        
+         if (\Auth::id() === $task->user_id) {
+            // メッセージ編集ビューでそれを表示
+            return view('tasks.edit', [
+                'task' => $task,
+            ]);
+         }
+         else {
+            // 前のURLへリダイレクトさせる
+            return back();
+        }
     }
 
     /**
@@ -121,15 +161,23 @@ class TasksController extends Controller
             'content' => 'required|max:255',
         ]);
         
-        //idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを更新
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
+        //  if (\Auth::id() === $task->user_id) {
+            //idの値でメッセージを検索して取得
+            $task = Task::findOrFail($id);
+            // メッセージを更新
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+            
+             // トップページへリダイレクトさせる
+            return redirect('/');
+        // } 
+        // else {
+        //     // 前のURLへリダイレクトさせる
+        //     return back();
+        // }
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+       
     }
 
     /**
@@ -143,9 +191,11 @@ class TasksController extends Controller
         //idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
         // メッセージを削除
-        $task->delete();
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+        }
 
-        // トップページへリダイレクトさせる
+        // 前のURLへリダイレクトさせる
         return redirect('/');
     }
 }
